+++ CHANGELOG.md
@@ -17,6 +17,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - TypeScript linting
 - Added Storybook Docs, enable to write MDX in separate file
 - Added custom Storybook theme
+- Added password component
 
 ### Changed